Move config to api directory
